feat: reuse text node as well
